Refactoring Update, Added directories and sub-directories.
