+++ Divination/DataOperations/AnalysisHelpers/cagr_calculator.py
