+++ Divination/parameters.py
